Add page-level padding and margin
The space on either side of the content is currently defined with
padding of <body> and <main>'s children.

This is not very responsive, but I'm unsure how to make it based off of
max-width as I'd prefer without refactoring some HTML. I either need to
discover some max-content-width CSS property, or I need to create
wrapper divs for each section to extend the padding. Unless there's some
wizardry with negative margins I could try?